move some utilities to a better location, plus a slight refactor
