+++ code/parse/sexp.h
@@ -1404,7 +1404,8 @@ extern void update_sexp_references(const char *old_name, const char *new_name, i
 extern std::pair<int, sexp_src> query_referenced_in_sexp(sexp_ref_type type, const char *name, int &node);
 extern void stuff_sexp_text_string(SCP_string &dest, int node, int mode);
 extern int build_sexp_string(SCP_string &accumulator, int cur_node, int level, int mode);
-extern int sexp_query_type_match(int opf, int opr);
+extern bool sexp_query_type_match(int opf, int opr);
+extern int sexp_match_closest_operator(const SCP_string &str, int opf);
 extern bool sexp_recoverable_error(int num);
 extern const char *sexp_error_message(int num);
 extern int count_free_sexp_nodes();
